--- .gitignore
@@ -1,4 +1,0 @@
-.pio.cache/
-npm-debug.log
-node_modules/
-/server/.pub/
