--- client/.gitignore
@@ -1,4 +1,0 @@
-/bootstrap.js
-/install.rdf
-/data/bundles/
-*.xpi
