--- client/.jpmignore
@@ -1,4 +1,0 @@
-/node_modules/pinf-for-mozilla-addon-sdk/node_modules/
-/node_modules/pinf-for-mozilla-addon-sdk/node_modules/**
-/node_modules/pinf-for-mozilla-addon-sdk/examples/
-/node_modules/pinf-for-mozilla-addon-sdk/examples/**
