--- client/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "firephp-for-firebug.next-client",
-  "title": "FirePHP for Firebug.Next",
-  "id": "firephp-for-firebug.next@firephp.org",
-  "description": "Log from server PHP scripts to the Firebug.Next console",
-  "author": "Christoph Dorn",
-  "license": "MPL 2.0",
-  "version": "0.1.0",
-  "main": "lib/main",
-  "dependencies": [
-    "pinf-for-mozilla-addon-sdk"
-  ]
-}
