+++ components/data/panel.html
